+++ src/main/java/io/redstudioragnarok/FBP/gui/menu/Page3.java
@@ -10,6 +10,8 @@ import net.minecraft.client.resources.I18n;
 
 import java.awt.*;
 
+import static io.redstudioragnarok.FBP.FBP.dynamicWeather;
+
 public class Page3 extends GuiScreen {
 
 	GuiButton b1, b2, b3, b4, b5, b6, Defaults, Done, Reload, Back, Next, Enable, ReportBug;
@@ -25,7 +27,7 @@ public class Page3 extends GuiScreen {
 		b1 = new FBPGuiButton(1, x, (this.height / 5) - 10 + GUIOffsetY, I18n.format("menu.fancyflame.info"), FBP.fancyFlame, true, true);
 		b2 = new FBPGuiButton(2, x, b1.y + b1.height + 1, I18n.format("menu.fancysmoke.info"), FBP.fancySmoke, true, true);
 		b3 = new FBPGuiButton(3, x, b2.y + b2.height + 6, I18n.format("menu.fancyweather.info"), FBP.fancyWeather, true, true);
-		b4 = new FBPGuiButton(4, x, b3.y + b3.height + 1, I18n.format("menu.fancysnow.info"), false, false, false);
+		b4 = new FBPGuiButton(4, x, b3.y + b3.height + 1, I18n.format("menu.dynamicWeather.title"), dynamicWeather, true, true);
 		b5 = new FBPGuiButton(5, x, b4.y + b4.height + 6, I18n.format("menu.waterphysics.info"), FBP.waterPhysics, true, true);
 		b6 = new FBPGuiButton(6, x, b5.y + b1.height + 1, I18n.format("menu.restonfloor.info"), false,false, false);
 
@@ -82,6 +84,9 @@ public class Page3 extends GuiScreen {
 		case 3:
 			FBP.fancyWeather = !FBP.fancyWeather;
 			break;
+		case 4:
+			dynamicWeather = !dynamicWeather;
+			break;
 		case 5:
 			FBP.waterPhysics = !FBP.waterPhysics;
 			break;
@@ -131,7 +136,7 @@ public class Page3 extends GuiScreen {
 					description = I18n.format("menu.fancyweather.description");
 					break;
 				case 4:
-					description = I18n.format("menu.fancysnow.description");
+					description = I18n.format("menu.dynamicWeather.description");
 					break;
 				case 5:
 					description = I18n.format("menu.waterphysics.description");